Refactor file directory to match Marvel directory style
