+++ build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'java'
     id 'org.springframework.boot' version '3.2.5'
     id 'io.spring.dependency-management' version '1.1.4'
+    id 'application'
 }
 
 group = 'com.example'
@@ -15,6 +16,14 @@ repositories {
     mavenCentral()
 }
 
+application {
+    mainClass = 'com.sg.gov.hdb.marvel.Demo2Application'  // Set your main class here
+}
+
+
+
+
+
 dependencies {
     implementation 'mysql:mysql-connector-java:8.0.28'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
@@ -26,6 +35,7 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.batch:spring-batch-test'
 }
+
 
 tasks.named('test') {
     useJUnitPlatform()
