+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'liquibase-poc-ud'
+rootProject.name = 'liquibase-poc-udy'
