+++ src/main/java/com/sg/gov/hdb/marvel/Demo2Application.java
@@ -1,10 +1,7 @@
-package com.integration;
+package com.sg.gov.hdb.marvel;
 
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.kafka.core.KafkaTemplate;
 
 @SpringBootApplication
 public class Demo2Application {
