+++ src/main/java/com/sg/gov/hdb/marvel/config/KafkaConfig.java
@@ -1,4 +1,4 @@
-package com.integration.config;
+package com.sg.gov.hdb.marvel.config;
 
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
