+++ src/main/java/com/sg/gov/hdb/marvel/service/KafkaProducerService.java
@@ -1,11 +1,11 @@
-package com.integration.service;
+package com.sg.gov.hdb.marvel.service;
 
+import com.sg.gov.hdb.marvel.model.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
-import com.integration.model.Message;
-import com.integration.repository.MessageRepository;
+import com.sg.gov.hdb.marvel.repository.MessageRepository;
 
 @Service
 public class KafkaProducerService {
