+++ src/test/java/com/sg/gov/hdb/marvel/Demo2ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.integration;
+package com.sg.gov.hdb.marvel;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
